+++ NEWS.md
@@ -3,12 +3,19 @@
 * Bug fixes:
   - `samp` works with size > length.
 
+* New functions
+  - `provide.file` To ensure a file (including its directory) exists.
+  
+* Enhancements:
+  - `provide.dir` returns the directory requested on success.
+
 # hutils 1.4.0
 * New functions:
   - `%<->%` To swap values between objects
   - `average_bearing`, the bearing bisecting two vectors
   - `dir2`, (Windows only) a much faster version of `dir()`
   - `Mode`, statistical mode
+
   - `replace_pattern_in` to find-and-replace on a pattern in all files in a directory
   - `samp`, a 'safe' version of `sample`.